+++ src/views/Dashboard.vue
@@ -21,53 +21,11 @@
         <span>Collection</span>
         <span class="material-icons"> library_books </span>
       </v-btn>
+      <v-btn value="profile" @click="changeRoute('profile')">
+        <span>Profile</span>
+        <span class="material-icons"> account_circle </span>
+      </v-btn>
     </v-bottom-navigation>
-
-    <!-- Desktop navigation drawer -->
-    <v-navigation-drawer permanent expand-on-hover class="nav-side">
-      <v-list>
-        <v-list-item class="px-2">
-          <v-list-item-avatar @click="changeRoute('profile')">
-            <v-avatar color="green" size="36">
-              <span class="black--text headline">{{ this.getAvatarName }}</span>
-            </v-avatar>
-          </v-list-item-avatar>
-        </v-list-item>
-
-        <v-list-item link @click="changeRoute('profile')">
-          <v-list-item-content>
-            <v-list-item-subtitle>{{ this.getUserEmail }}</v-list-item-subtitle>
-          </v-list-item-content>
-        </v-list-item>
-      </v-list>
-
-      <v-divider></v-divider>
-
-      <v-list nav dense>
-        <v-list-item link @click="changeRoute('main')">
-          <v-list-item-icon>
-            <span class="material-icons"> home </span>
-          </v-list-item-icon>
-          <v-list-item-title>Dashboard</v-list-item-title>
-        </v-list-item>
-        <v-list-item link @click="changeRoute('review')">
-          <v-list-item-icon>
-            <span class="material-icons"> alarm </span>
-          </v-list-item-icon>
-          <v-list-item-title>Review</v-list-item-title>
-        </v-list-item>
-        <v-list-item link @click="changeRoute('library')">
-          <v-list-item-icon>
-            <span class="material-icons"> library_books </span>
-          </v-list-item-icon>
-          <v-list-item-title>Library</v-list-item-title>
-        </v-list-item>
-      </v-list>
-
-      <v-list>
-        <v-list-item> </v-list-item>
-      </v-list>
-    </v-navigation-drawer>
   </section>
 </template>
 
@@ -75,18 +33,6 @@
 import firebase from "firebase";
 
 export default {
-  data: () => ({}),
-  computed: {
-    getAvatarName() {
-      return this.$store.getters.getAvatarName;
-    },
-    getUserId() {
-      return this.$store.getters.getUserInfo.id;
-    },
-    getUserEmail() {
-      return firebase.auth().currentUser.email;
-    },
-  },
   methods: {
     changeRoute(rn) {
       if (this.$route.name != rn) {
@@ -105,36 +51,15 @@ export default {
 </script>
 
 <style>
+.dashboard-container {
+  padding: 0.5rem 1rem;
+}
+
 .v-avatar > span {
   cursor: pointer;
 }
 
-.content {
-  margin-bottom: 4rem;
-}
-
-.nav-side {
-  display: none !important;
-}
-
 .nav-bottom {
   height: 37px !important;
-}
-
-@media only screen and (min-width: 1024px) {
-  .dashboard-container {
-    max-width: 90vw !important;
-  }
-
-  .nav-bottom {
-    display: none !important;
-  }
-
-  .nav-side {
-    z-index: 4;
-    position: fixed;
-    left: 0;
-    display: block !important;
-  }
 }
 </style>
